Fix 5555408: allow emergency button to be disabled on SIM/PUK unlock screen
Some carriers cannot process emergency calls when the IMSI is not available,
which happens when the SIM is locked.  This allows the behavior to be overwritten
in an overlay. When the config variable is set to 'false', the emergency
button no longer appears on the SIM or PUK unlock screens.

Change-Id: I3127bfd5e7ebfabea997fbe048cc474e48b7a81c

Former-commit-id: 1f56edc6cd377eab05d9816e93591a7139cdff96