NDK: ANativeActivity: Fix documentation for 'clazz' member.
As discussed on the android-ndk forum, it turns out that the
member named 'clazz' is really a global reference to the NativeActivity
instance. As such, it's really a handle to a VM object, not a VM class.

Ideally, we would rename it to 'activity', but this cannot be done
without breaking NDK source compatibility.

Change-Id: I82ca1549b35346a3eacf9e84c4c836387fc883a6

Former-commit-id: dd00e734f335f13b4a9e24a27e280b6112d919a5