Fix bug 5544103 - Spinner text doesn't look dimmed when disabled
Pre-Holo spinners had a far more "buttony" look that expressed the
disabled state clearly, but Holo spinners are more subtle. As a result
we want to mark the contained view representing the current item as
disabled when the spinner itself is disabled.

Express this as a private framework style attribute that cannot be
changed at runtime.

Change-Id: Icff2ef2b8a3b1a96cbf00e4c75eda41a4dada7b3

Former-commit-id: 42b7e99b11a5ab1cbc0beebe0b15e46bdf462dff