Bring back the 'more notifications' icon.
In fact, there's even more going on here. Icons were
appearing cut off because of the unusual width of icons on
the right-hand side of the status bar; we now take extra
steps to ensure that the IconMerger is only wide enough to
show an integral number of icons. But that causes problems
with the rest of the status bar the IconMerger makes itself
thinner than the overall layout requires, so there's now a
containing layout to absorb the additional horizontal space
properly.

Bug: 5136545

Change-Id: I8c4d1ee2bcd976c452510b33c196fbe0c3e8c660

Former-commit-id: 05e2414ef5a3a8fecbabc1f220c9925539d51f7a