Fix 5602985: Show "Emergency calls only" when device has no SIM.
Some countries require phones to show "Emergency calls only" when
roaming on a device with no Sim.  This changes causes the phone
to now display the PLMN string on devices that are voice-capable when
the sim is missing.  This depends on the PLMN containing the text
"Emergency calls only" when the radio has connectivity to any carrier.
Otherwise, the PLMN is expected be null or empty which will cause the
device to just show "No SIM card."

Change-Id: I7c1d4624ae67074e2ca40161845ca8bb36031c42

Former-commit-id: 2fcd63b8cf94a947498f0aab24b9d702a6fe909b