Quick fix for bug 5646217 blocking automation tests
The screen saver knows to avoid running when applications
are holding wake locks via window flags, but other wake
locks are invisible to the window manager and therefore to
the screen saver activation code.

The correct thing to do is move the screen saver to the
power manager to give it access to all wakelock information.
In the meantime, this fix checks for the user's
STAY_ON_WHILE_PLUGGED_IN setting and disables the screen
saver in this case.

Bug: 5646217
Change-Id: Ia2286a8c3745267871901f7a0b7df94ec344a456

Former-commit-id: 7d46c613f6bfeb6638c6aafb52eab417ad8eef52