Swap OK and Cancel buttons in SIM PIN unlock screen to match ICS guidelines.
The ICS navigation guidelines state that the OK action should be to the
right of the cancel action in UI dialogs. Update the SIM PIN unlock and
SIM PUK unlock layouts to swap the OK and Cancel buttons (for soft keypad)
and the OK and Emergency Call buttons for the physical keyboard landscape
layout.

Bug: 5730977
Change-Id: Iaf565bc2b584292a34f5a72a7a763003f9149f27

Former-commit-id: 48d50958af05db84ea787f5e4096ff40307ec804