Text blinking cursor hard to see.
Bug 5738416

This problem was introduced in
https://android-git.corp.google.com/g/#/c/152599

The invalidated cursor region is no longer expanded to include the
cursor drawable (different from the handles' drawables).

Added that code back. Added a flag, since invalidateRegion is also
used to invalidate text span regions in spell check (assumes the decorated
span bounds are not bigger that text boundaries), which is fine in case
of underline.

When the cursor is moved on a single line, invalidate the whole line
(same as what was done before). This is sub-optimal, will file a bug
to fix this. The core problem is that we should invalidate the previous
and the new cursor's positions. We only have one of these.

Change-Id: I9ada9340fb52aad3d80c39efd021fd3f9ec0cc4d

Former-commit-id: 961ebb9ab0a6d45b06a74aa90894f7fda3d528c6