Changed how Face Unlock coordinates are specified
Was using View.getLeft() and View.getTop() to specify the upper-left
corner of the Face Unlock area.  That gives coordinates relative the
view, which was fine for the phones.  For the tablet it needs
coordinates relative to the window (which still works for the phones).

Also fixed a 'bug' where h and w were swapped.  However, it wasn't
causing a problem because it was swapped in two places.

Change-Id: I86c1f68439f1dcef826cfe6b8fb56c9a4a6b8dc3

Former-commit-id: a44f2a59a2408ff367bd2efc7f74ec56f92565dc