Fix issue with selecting proper JNI libraries
Current code will load both abi and abi2 versions of the JNI libraries.
The second one found will be the library in use. Update algorithm so
that abi2 libraries will never be loaded once we find a library that
supports the primary abi. This is consistent with the original
implementation.

Bug: 5787224
Change-Id: Ib4d30bea8316ca9bf62757c970dcc5b75cd969ec

Former-commit-id: 10410fc85248ea77c2c1fbbb666c3fe4fcfef1b6