AOOB in AutoComplete
Bug 5749557

Not clear how we can get an AOOB in that case.
tmp will always have the right length, and indeed the stack
trace attached to that bug shows a correct size of 10.

However, there is an index issue when we build the new
completion array. i is not the correct index to use.

Note however that the original buildDropDown method mentioned
is no longer present in the file. I tried to backtrack, but
the use of arraycopy always seemed correct.

Change-Id: Idf749c74b38923b5d18596c8e8f6ea887cc897d6

Former-commit-id: f0d3b7ff72511c9f6b9bf228e5b14ba6a2e35f34