Detect and repair invalid layouts in the navigation bar.
It seems that something (probably layout transitions)
occasionally leaves the navigation bar layout in a weird
state such that most of the buttons are offscreen. For
example, the portrait nav bar might have buttons with widths
matching the nav bar's old width from landscape; the
clickable regions will therefore be far off the right edge
of the display.

The navigation bar now looks for this situation when the
size of the view changes and corrects for it.

Bug: 5549288
Change-Id: I8464ac9fd64cb64ec18baf51478dd0f6564a2284

Former-commit-id: 0b69b59877896e4cab448cef9e12ff4c29a4252e