Remove @deprecate in View's fadingEdge
The comment of this field is also aggregated in the android.R.styleable#View
description. As a result, the @deprecated in the comment of this field also
applies to android.R.styleable#View which incorrectly appears deprecated.

This fixes the problem, although fadingEdge will no longer be marked as
deprecated in IDEs. I believe it is less important that to make View
deprecated.

Change-Id: I96d8e868001c4c853eb1536f8401275b9b1e689b

Former-commit-id: 259c1090464b966b8019abfae848a96895ecdbcd