EditText-specific data extracted from TextView
All edit-specific data has been moved to an EditData inner class in TextView.

The instance of this class is created as soon as one of its fields is set to
a non-default value. Regular TextViews (buttons, checkboxes...) should never
have to instantiate the Editor, thus reducing their memory footprint.

Note: this is a debug version. The getEditor() method is here to track
possible problems and should be removed in the final version. Indentation
will be fixed then.

Next step is to extract more methods and classes into Editor and create a
dedicated class for it. mMovement may also be Editor specific.

Change-Id: Ic241953a2fb73213180f300c3609a9e6799aa884

Former-commit-id: 60e2186c354c30b2f75ed0ef1ba75181bd32afda