Add Developer Option setting for Animator scaling.
This new setting allows users to set a scale factor for the
duration and startDelay of all Animator-based animations. This
setting is very similar to the Transition animation scale and
Window animation scale settings, except this one applies specifically
to Animator animations. The property is only accessible by users
through the Settings UI, not programmatically. The value applies
system-wide and is picked up per-process at the time of the first
ValueAnimator construction.

Change-Id: I3d5fbc956695c88d01c30820259da3e107ffd8a3

Former-commit-id: d21a9fe2d9a6cfe966fc7df3a8c37c172d7ac302