Make aapt ignore tools-related data.
This patchset introduces a new standard namespace http://schemas.android.com/tools
which will be used for tools specific XML attributes.
Any attributes using this namespace will not be compiled into the binary XML file.

The namespace node is also not written at all, and its string is not collected
to ensure that there is no impact on the devices.

Change-Id: I3e75d44cda54e1fa7b5cdc56b3eb27db80fe7761

Former-commit-id: a5d5e9d906583ebeb590ca200ca9840302a5cd1a