audio preprocessing: fix multiple enable problem
Do not set the effect state back to configured when a configuration
command is received while the effect is enabled. Instead just check that
the new config is the same as current config.

It is normal to receive a config command from the framework while enabled
if a new effect is added on the same session.

Change-Id: I93aa38bf60a3d7cc7729934e87ddd69bf1112cd6

Former-commit-id: 8293e19d47376c3fb2444c942275a2876f5763c3