Simplify Youtube <embed> management
When we encounter youtube <embed> objects, we replace them on the fly
with an element that will redirect to the youtube application.
Currently, this element is a canvas, that we add to the document wrapped
into a Frame. This cause some problems when we switch to have canvas
elements on their own individual layer (which drastically improves
performances). Note that we do support correctly composited layers
into frameset/iframes on normal websites, the problems we see here
are due to us creating the frame on the fly, messing with webkit's
compositing logic.
This CL rewrite the HTML code we insert to not use canvas and instead
use normal html elements positioned via CSS. This work around the
composited canvas issue as well as simplifying the code.

Change-Id: Ie6043f9445e8bc191b229db9f9ff5de192d8b5db

Former-commit-id: 4d49c7f04ab34686de0da5aba6f7427a151c471d