Selection highlight made possible with no editor
Bug 6065081

Selecting text programatically is possible (like when a link is highlighted
by LinkMovementMethod) even when the text is not editable.

Also removed the textCanBeSelected text is highlight computation which prevented
links from being highlighted (legacy problem, not reported).
Now that they work, we can see that the highlight is not removed when the text
loses focus. Will be addressed in a different CL.

Change-Id: I673f435966686234b4c0b0a97bcc65abd01169ff

Former-commit-id: 83051b8c75a145a12eca1a021ff1a4fbcc77ba4b