Don't close the database until all references released.
SQLiteDatabase.close() should call releaseReference() rather than
closing the database immediately.  SQLiteDatabase should also hold
a reference to itself while performing certain operations to
ensure that they complete normally even if another thread closes
the database at the same time.

Fixed a couple of missing or redundant uses of acquireReference()
related to CursorWindows.

To be honest, the reference counting performed by SQLiteClosable should
not be needed, but we're stuck with it in the API.

Bug: 6104842
Change-Id: I3444a697409905d4a36b56418dc7766f5ba76b59

Former-commit-id: 03bd302aebbb77f4f95789a269c8a5463ac5a840