scroll relative to visibleRect in spawnContentScrollTo()
WebKit calculates its scrolling destination coordinates relative to the
visibleRect that WebCoreViewBridge provides it. That visibleRect is stored
in WebViewClassic as mScrollOffset. When a titlebar is displaying, the
coordinate conversion from content to view fails. This conversion is used
in many other places and seems to function perfectly fine for them, so I've
concluded that while it may be the wrong tool for this particular job, it is
the correct tool for many others. As a result, I've left the conversion as it
is and simply fixed the function that WebKit calls to programatically scroll
the content. This fixes all of the programmatic scrolling issues I've been
seeing with find-on-page without breaking anything else.

  Bug: 5470588

Change-Id: I50d3af4dd8a7fbd2d04bbb41e38f3e6947fbb46a

Former-commit-id: 16e25e6ad8659f8b22f4dc0cb9f6ea95a9e7493f